src/OryxESS.Web1/OryxESS-Web/src/app/iou/+iouheader/iou-form2.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { IouForm2Component } from './iou-form2.component';

describe('Component: IouForm2', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [IouForm2Component]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([IouForm2Component],
      (component: IouForm2Component) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(IouForm2ComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(IouForm2Component));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-iou-form2></app-iou-form2>
  `,
  directives: [IouForm2Component]
})
class IouForm2ComponentTestController {
}

